Ensure we handle exceptions while handling unhandled exceptions
When an unhandled exception occurs, we try and be nice and display it.
If an exception occurs while doing *that*, swallow (and log it). This can
happen in some edge cases (it used to happen when shutting down, but that
was fixed), and if we throw the derived exception we'll put the wrong
stack trace into the event log.

Helps with #15