Always use HTTPS
There's a high cost involved with using http when https is required
(i.e. a redirect on each request), and a small cost with using https
when we can get away with http (a TLS handshake).

It's good practice to always use https for the GUI / REST API, and
SyncTrayzor makes this transparent to the user (they won't get a
certificate warning in SyncTrayzor). Therefore it makes a lot of sense
to mandate it.

Fixes #33