Ensure the API client is created, and initial data loaded, *before* starting pollers
We used to use the StartupComplete event as our indicator that Syncthing's
API was working. Since we now go through the API client creation phase,
this is no longer necessary. Therefore we count Syncthing as having
started once we managed to get a single REST connection to it. Then we
load the initial data, and finally kick off the event watcher.

squash! Ensure the API client is create, and initial data loaded, *before* starting pollers