Fix oversize issue on Settings screen re folders list
If there are too many folders, the Settings screen would happily size
itself beyond the bounds of the window. Not ideal.

I've gone for making the folders list scrollable, and automatically setting the
window's MaxHeight so that it doesn't push the save/cancel buttons below
the bottom border of the window.

There's potential confusion here because we don't allow the user to resize
the Settings screen, but that's a significant headache in itself

Fixes #76