Drop Syncthing 0.10, and don't replace binary in AppData
There's more thinking to be done around when we replace the Syncthing
binary in AppData, but it's not required right now.

This fixes a potential issue where people let Syncthing upgrade itself
to 0.11, but the installer's still in 0.10 mode, so their 0.11 version
gets overwritten by the installer. Not ideal.

We now only write the binary to AppData if it doesn't already exist.
This puts more responsibility on the user to keep this up to date,
but it should be auto-updated by default. In future we may want some
mechanism to run 'syncthing.exe -version' and check whether it's too
old: if it is, offer to replace it. However, there's no current
(known) 'too old to be compatible' version, so this isn't (yet) an
issue

Fixes #66