Allow multiple auto-started SyncTrayzor installations
Previously, SyncTryazor assumed there was a single installation, and so
a single auto-start registry key. If there *were* multiple installations,
then they would all try and manipulate the same registry key, including
automatically updating the key to point to itself on startup.

Now, on startup SyncTrayzor will try and find a key called 'SyncTrayzor',
'SyncTrayzor (2)', etc, which points to its current location. If found,
this is the key that is manipulated. If not, then a new name is generated
(according to the same numbering scheme)

Fixes #81