Assume all in-progress file transfers complete when folder finishes syncing
Syncthing may not tell us that a file transfer completed, because it forgot
about the event (it only caches 1000 events, and we simply can't fetch them
quickly enough).

If this happens, we'll end up with files that are still 'starting' when
the folder finishes synchronizing. In this case, assume those files finished.

Relates to #101, #106