Be much more graceful when restarting for an update
We used to taskkill /f /t /im SyncTrayzor, as that was the only way to
get CefSharp to stop. In 1.0.18 we added support for the Restart Manager,
and gracefully shut down CefSharp and syncthing in this case.

Now enough time has passed by that we can alter the installer to take
advantage of this. It now uses the Restart Manager to stop SyncTrayzor,
which stops everything else in a nice clean way.

Anyone pre-1.0.18 who's upgrading will probably see the 'could not
stop...' error message, but that's their fault for not keeping with the
times :)