Base 'failed folders' on FolderErrors event
It turns out it's possible for a file transfer to succeed and fail at the
same time. Consider the case where Syncthing couldn't delete a folder
because it has an ignored file in it. Then the user deletes the folder from
the filesystem. Now Syncthing attempts to delete the folder and fails
because it no longer exists, but we're now in sync because the folder no
longer exists. We won't get a succeeded transfer for the folder, but we're
in sync.