Don't spam connected/disconnected events if they happen quickly
This can happen if a remote device has us paused, for example. Syncthing may
modify BEP to give us more information in the Connected event in the future
(syncthing/syncthing#3379), but for now just delay the Connected event
until we're sure that it isn't followed straight away by a Disconnected
event.

Fixes #289