Add the --start-syncthing and --stop-syncthing command-line flags
If SyncTrayzor is already running, these tell the running instance to
start / stop Syncthing, respectively. If it isn't, these override the
config setting which says whether to automatically start Syncthing.

Both of these imply --minimized: that is, they do not cause SyncTrayzor's
main window to be shown. To override this, a new flag --show has been
added.

Fixes #350