Make FreePortFinder more resilient
User reported that FreePortFinder was failing to find a free port when
another instance of Syncthing was running - it was rethrowing the exception
after the first attempt, which means it's a SocketError I wasn't anticipating.
There isn't enough information to say what SocketError it was though.

Just continue on all of them. We'll give up after a few anyway.

Fixes #381