Don't store an API Key in config
There's no need to any more. Just generate a new one on each startup.

We used to generate and store one in the days when Syncthing could only
accept a single API key. Therefore we let the user change the API key
in our config, so they could also use that AIP key in other applications.
Since Syncthing began allowing us to specify an additional API key, this
became no longer necessary. We remove the API key from the settings UI,
but did not remove it from the config.